StdLibDevConsole: a replacement for StdLib/LibC/Uefi/Devices/Console
This is for linking with "POSIX" applications. If EFI ConOut is graphical,
but StdErr is redirected to a non-graphical console, stderr will always
send to ConOut, not StdErr. This solves the unexpected
behavior of POSIX apps seemingly printing nothing on error on
certain UEFI implementations that always redirect StdErr to serial.

Signed-off-by: Andrei Warkentin <andrey.warkentin@gmail.com>